requirements.txt
requests
python-dotenv
streamlit
langchain-community
asyncio
crawl4ai
snowflake
snowflake-connector-python
pathlib
firebase_admin
os
python-dotenv


